+++ packages/frontend/src/features/layout/hooks/use-sidebar.ts
@@ -2,7 +2,7 @@ import { useLocalStore } from "@/context/local-state";
 
 export const useSidebar = () => {
   const width = useLocalStore((state) => state.sidebarWidth);
-  const setWidth = useLocalStore((state) => state.setCalculatedWidth);
+  const setWidth = useLocalStore((state) => state.setSidebarWidth);
   const isSidebarCollapsed = useLocalStore((state) => state.isSidebarCollapsed);
   const setSidebarCollapsed = useLocalStore(
     (state) => state.setSidebarCollapsed,